Always delete lightsheet coordinates in finalize
For caustics the lightsheet coordinates are accessible as UV-layers, this can
be used to display a distorted image as if it was projected from the lamp. But
setting up such a scene is complicated. It does not look good in Cycles (if the
caustic gets folded) and in EEVEE you cannot setup a lamp as a projector
anyway.
This commit removes one setting from the finalize operator, now it will always
delete the UV-layers. If you still want to use this projection effect, just
don't finalize your caustics!