Make caustics visible in EEVEE by default
An API change extended the visibility settings for objects, the new settings
made caustics invisible. This is correct for Cycles (only diffuse is visible),
but not for EEVEE (and probably not for other render engines).
Now we disable camera visibility if Cycles is the selected render engine. This
is only checked when the caustic is created.